+++ .dockerignore
@@ -1,0 +1,4 @@
+venv
+__pycache__
+*/*/__pycache__
+*/*/venv
