+++ .gitignore
@@ -1,0 +1,8 @@
+__pycache__/
+venv/
+
+# Generated by review action
+clang-tidy-review-output.json
+
+# Generated by clang-tidy
+clang_tidy_review.yaml
