+++ action.yml
@@ -53,8 +53,12 @@ inputs:
     description: 'Message to post on PR if no issues are found. An empty string will post no LGTM comment.'
     required: false
     default: 'clang-tidy review says "All clean, LGTM! :+1:"'
+  post_comments:
+    description: 'Post comments and review about any issues found. If set to false, can be accompanied by the post_comments child action'
+    required: false
+    default: true
   pr:
-    default: ${{ github.event.pull_request.number }}
+    default: ${{ github.event.number }}
   repo:
     default: ${{ github.repository }}
 outputs:
@@ -77,3 +81,4 @@ runs:
     - --apt-packages=${{ inputs.apt_packages }}
     - --cmake-command='${{ inputs.cmake_command }}'
     - --lgtm-comment-body='${{ inputs.lgtm_comment_body }}'
+    - --post_comments=${{ inputs.post_comments }}
